+++ agent_lib/CMakeLists.txt
@@ -246,13 +246,13 @@ set(AGENT_SOURCES
 
 # src/sink/mqtt_sink HEADER_FILE_ONLY
 
-        "${SOURCE_DIR}/sink/mqtt_sink/mqtt_service.hpp"
-	"${SOURCE_DIR}/sink/mqtt_sink/mqtt2_service.hpp"
+        "${SOURCE_DIR}/sink/mqtt_sink/mqtt_legacy_service.hpp"
+	"${SOURCE_DIR}/sink/mqtt_sink/mqtt_service.hpp"
 
 #src/sink/mqtt_sink SOURCE_FILES_ONLY
 
-        "${SOURCE_DIR}/sink/mqtt_sink/mqtt_service.cpp"
-	"${SOURCE_DIR}/sink/mqtt_sink/mqtt2_service.cpp"
+        "${SOURCE_DIR}/sink/mqtt_sink/mqtt_legacy_service.cpp"
+	"${SOURCE_DIR}/sink/mqtt_sink/mqtt_service.cpp"
         
 # src/sink/rest_sink HEADER_FILE_ONLY
