+++ test_package/CMakeLists.txt
@@ -249,8 +249,8 @@ add_agent_test(tls_http_server FALSE sink/rest_sink TRUE)
 add_agent_test(routing FALSE sink/rest_sink)
 
 add_agent_test(mqtt_isolated FALSE mqtt_isolated TRUE)
+add_agent_test(mqtt_legacy_sink FALSE sink/mqtt_legacy_sink TRUE)
 add_agent_test(mqtt_sink FALSE sink/mqtt_sink TRUE)
-add_agent_test(mqtt_sink_2 FALSE sink/mqtt_sink_2 TRUE)
 
 add_agent_test(json_printer_asset TRUE json)
 add_agent_test(json_printer_error TRUE json)
