+++ test_package/agent_test_helper.hpp
@@ -29,7 +29,7 @@
 #include "mtconnect/configuration/agent_config.hpp"
 #include "mtconnect/configuration/config_options.hpp"
 #include "mtconnect/pipeline/pipeline.hpp"
-#include "mtconnect/sink/mqtt_sink/mqtt2_service.hpp"
+#include "mtconnect/sink/mqtt_sink/mqtt_legacy_service.hpp"
 #include "mtconnect/sink/mqtt_sink/mqtt_service.hpp"
 #include "mtconnect/sink/rest_sink/response.hpp"
 #include "mtconnect/sink/rest_sink/rest_service.hpp"
@@ -124,8 +124,8 @@ public:
 
   ~AgentTestHelper()
   {
+    m_mqttLegacyService.reset();
     m_mqttService.reset();
-    m_mqtt2Service.reset();
     m_restService.reset();
     m_adapter.reset();
     if (m_agent)
@@ -171,21 +171,21 @@ public:
     return rest;
   }
 
-  std::shared_ptr<mtconnect::sink::mqtt_sink::MqttService> getMqttService()
+  std::shared_ptr<mtconnect::sink::mqtt_sink::MqttLegacyService> getMqttLegacyService()
   {
     using namespace mtconnect;
-    sink::SinkPtr sink = m_agent->findSink("MqttService");
-    std::shared_ptr<mtconnect::sink::mqtt_sink::MqttService> mqtt =
-        std::dynamic_pointer_cast<mtconnect::sink::mqtt_sink::MqttService>(sink);
+    sink::SinkPtr sink = m_agent->findSink("MqttLegacyService");
+    std::shared_ptr<mtconnect::sink::mqtt_sink::MqttLegacyService> mqtt =
+        std::dynamic_pointer_cast<mtconnect::sink::mqtt_sink::MqttLegacyService>(sink);
     return mqtt;
   }
 
-  std::shared_ptr<mtconnect::sink::mqtt_sink::Mqtt2Service> getMqtt2Service()
+  std::shared_ptr<mtconnect::sink::mqtt_sink::MqttService> getMqttService()
   {
     using namespace mtconnect;
-    sink::SinkPtr mqttSink = m_agent->findSink("Mqtt2Service");
-    std::shared_ptr<mtconnect::sink::mqtt_sink::Mqtt2Service> mqtt2 =
-        std::dynamic_pointer_cast<mtconnect::sink::mqtt_sink::Mqtt2Service>(mqttSink);
+    sink::SinkPtr mqttSink = m_agent->findSink("MqttService");
+    std::shared_ptr<mtconnect::sink::mqtt_sink::MqttService> mqtt2 =
+        std::dynamic_pointer_cast<mtconnect::sink::mqtt_sink::MqttService>(mqttSink);
     return mqtt2;
   }
 
@@ -199,8 +199,8 @@ public:
     using ptree = boost::property_tree::ptree;
 
     sink::rest_sink::RestService::registerFactory(m_sinkFactory);
+    sink::mqtt_sink::MqttLegacyService::registerFactory(m_sinkFactory);
     sink::mqtt_sink::MqttService::registerFactory(m_sinkFactory);
-    sink::mqtt_sink::Mqtt2Service::registerFactory(m_sinkFactory);
     source::adapter::shdr::ShdrAdapter::registerFactory(m_sourceFactory);
 
     ConfigOptions options = ops;
@@ -232,24 +232,24 @@ public:
     m_restService = std::dynamic_pointer_cast<sink::rest_sink::RestService>(sink);
     m_agent->addSink(m_restService);
 
+    if (HasOption(options, "MqttLegacySink"))
+    {
+      auto mqttContract = m_agent->makeSinkContract();
+      mqttContract->m_pipelineContext = m_context;
+      auto mqttsink = m_sinkFactory.make("MqttLegacyService", "MqttLegacyService", m_ioContext,
+                                         std::move(mqttContract), options, ptree {});
+      m_mqttLegacyService = std::dynamic_pointer_cast<sink::mqtt_sink::MqttLegacyService>(mqttsink);
+      m_agent->addSink(m_mqttLegacyService);
+    }
+
     if (HasOption(options, "MqttSink"))
     {
       auto mqttContract = m_agent->makeSinkContract();
       mqttContract->m_pipelineContext = m_context;
-      auto mqttsink = m_sinkFactory.make("MqttService", "MqttService", m_ioContext,
-                                         std::move(mqttContract), options, ptree {});
-      m_mqttService = std::dynamic_pointer_cast<sink::mqtt_sink::MqttService>(mqttsink);
-      m_agent->addSink(m_mqttService);
-    }
-
-    if (HasOption(options, "Mqtt2Sink"))
-    {
-      auto mqttContract = m_agent->makeSinkContract();
-      mqttContract->m_pipelineContext = m_context;
-      auto mqtt2sink = m_sinkFactory.make("Mqtt2Service", "Mqtt2Service", m_ioContext,
+      auto mqtt2sink = m_sinkFactory.make("MqttService", "MqttService", m_ioContext,
                                           std::move(mqttContract), options, ptree {});
-      m_mqtt2Service = std::dynamic_pointer_cast<sink::mqtt_sink::Mqtt2Service>(mqtt2sink);
-      m_agent->addSink(m_mqtt2Service);
+      m_mqttService = std::dynamic_pointer_cast<sink::mqtt_sink::MqttService>(mqtt2sink);
+      m_agent->addSink(m_mqttService);
     }
 
     m_agent->initialize(m_context);
@@ -324,8 +324,8 @@ public:
   mhttp::Server *m_server {nullptr};
   std::shared_ptr<mtconnect::pipeline::PipelineContext> m_context;
   std::shared_ptr<adpt::shdr::ShdrAdapter> m_adapter;
+  std::shared_ptr<mtconnect::sink::mqtt_sink::MqttLegacyService> m_mqttLegacyService;
   std::shared_ptr<mtconnect::sink::mqtt_sink::MqttService> m_mqttService;
-  std::shared_ptr<mtconnect::sink::mqtt_sink::Mqtt2Service> m_mqtt2Service;
   std::shared_ptr<mtconnect::sink::rest_sink::RestService> m_restService;
   std::shared_ptr<mtconnect::source::LoopbackSource> m_loopback;
